+++ src/batou/__init__.py
@@ -1,11 +1,19 @@
 # This code must not cause non-stdlib imports to support self-bootstrapping.
 import os.path
+import os
 import traceback
 
 from ._output import output
 
 with open(os.path.dirname(__file__) + "/version.txt") as f:
     __version__ = f.read().strip()
+
+# Configure `remote-pdb` to be used with `breakpoint()` in Python 3.7+:
+os.environ['PYTHONBREAKPOINT'] = "remote_pdb.set_trace"
+if not os.environ.get('REMOTE_PDB_HOST', None):
+    os.environ['REMOTE_PDB_HOST'] = "127.0.0.1"
+if not os.environ.get('REMOTE_PDB_PORT', None):
+    os.environ['REMOTE_PDB_PORT'] = "4444"
 
 
 class ReportingException(Exception):